fix: standardize requests for events, dates and categories
