+++ modules/@frontend_organizer-app/src/components/Fetchers/CategoriesFetcher/locales.ts
@@ -1,0 +1,9 @@
+import i18n from '@frontend/core/lib/utils/i18n';
+
+i18n.addResourceBundle('en', 'categories_fetcher', {
+    categories_fetch_error: 'Unable to fetch categories',
+});
+
+i18n.addResourceBundle('fr', 'categories_fetcher', {
+    categories_fetch_error: 'Impossible de récuperer les catégories',
+});
