+++ modules/@frontend_organizer-app/src/components/Fetchers/DatesFetcher/locales.ts
@@ -1,0 +1,9 @@
+import i18n from '@frontend/core/lib/utils/i18n';
+
+i18n.addResourceBundle('en', 'dates_fetcher', {
+    dates_fetch_error: 'Unable to fetch dates',
+});
+
+i18n.addResourceBundle('fr', 'dates_fetcher', {
+    dates_fetch_error: 'Impossible de récuperer les dates',
+});
