+++ modules/@frontend_organizer-app/src/components/Fetchers/EventsFetcher/locales.ts
@@ -1,0 +1,9 @@
+import i18n from '@frontend/core/lib/utils/i18n';
+
+i18n.addResourceBundle('en', 'events_fetcher', {
+    events_fetch_error: 'Unable to fetch events',
+});
+
+i18n.addResourceBundle('fr', 'events_fetcher', {
+    events_fetch_error: 'Impossible de récuperer les évènements',
+});
