+++ modules/@frontend_organizer-app/src/screens/CreateEvent/index.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from 'react';
+import React, { useContext, useEffect, useState } from 'react';
 import styled                                           from 'styled-components';
 import { useDispatch } from 'react-redux';
 
@@ -28,6 +28,7 @@ import { useToken } from '@frontend/core/lib/hooks/useToken';
 import { b64ImgtoBlob } from '../../utils/b64ImgtoBlob';
 import { useUploadImage } from '@frontend/core/lib/hooks/useUploadImage';
 import { uploadImageWithSdk } from '../../utils/uploadImageWithSdk';
+import { EventsContext } from '@frontend/core/lib/contexts/EventsContext';
 
 export interface FormProps {
     onComplete: () => void;
@@ -82,10 +83,10 @@ const updateCategories = (categories: CategoryWithDatesPayload[]): CategoryWithD
         currency: category.currency.toUpperCase()
     }));
 
-
 const CreateEvent: React.FC = () => {
     const [ t ] = useTranslation('create_event');
 
+    const eventReq = useContext(EventsContext);
     const [ currentStep, setCurrentStep ] = useState<number>(0);
 
     const [redirecting, setRedirecting] = useState<boolean>(false);
@@ -217,6 +218,7 @@ const CreateEvent: React.FC = () => {
             formik.resetForm();
             formik.validateForm();
             setRedirecting(true);
+            eventReq.force();
             setTimeout(() => {
                 localStorage.removeItem('event-creation');
                 history.push('/');
