+++ modules/@frontend_organizer-app/src/screens/Dashboards/CategoriesDashboard/locales.ts
@@ -4,16 +4,10 @@ i18n.addResourceBundle('en', 'categories_dashboard', {
     categories_of: 'Categories of',
     no_category_msg: 'There is no category for this event',
     create_category: 'Create category',
-    event_fetch_error: 'Unable to fetch event',
-    dates_fetch_error: 'Unable to fetch dates',
-    categories_fetch_error: 'Unable to fetch categories',
 });
 
 i18n.addResourceBundle('fr', 'categories_dashboard', {
     categories_of: 'Catégories de',
     no_category_msg: 'Il n\'y a aucune catégorie pour cet évènement',
     create_category: 'Créer une catégorie',
-    event_fetch_error: 'Impossible de récuperer l\'évènement',
-    dates_fetch_error: 'Impossible de récuperer les dates',
-    categories_fetch_error: 'Impossible de récuperer les catégories',
 });
