+++ modules/@frontend_organizer-app/src/screens/Dashboards/CategoriesDateDashboard/locales.ts
@@ -4,14 +4,10 @@ i18n.addResourceBundle('en', 'categories_date_dashboard', {
     categories_of: 'Categories of',
     no_category_msg: 'There is no category for this date',
     create_category: 'Create category',
-    dates_fetch_error: 'Unable to fetch dates',
-    categories_fetch_error: 'Unable to fetch categories',
 });
 
 i18n.addResourceBundle('fr', 'categories_date_dashboard', {
     categories_of: 'Catégories de',
     no_category_msg: 'Il n\'y a aucune catégorie pour cette date',
     create_category: 'Créer une catégorie',
-    dates_fetch_error: 'Impossible de récuperer les dates',
-    categories_fetch_error: 'Impossible de récuperer les catégories',
 });
