+++ modules/@frontend_organizer-app/src/screens/UpdateCategories/EditCategory/index.tsx
@@ -1,57 +1,19 @@
 import React from 'react';
 import { useParams } from 'react-router';
-import { useTranslation } from 'react-i18next';
-
-import { useRequest } from '@frontend/core/lib/hooks/useRequest';
-import { useToken } from '@frontend/core/lib/hooks/useToken';
-
-import { FullPageLoading, Error } from '@frontend/flib-react/lib/components';
-import { DatesSearchResponseDto } from '@common/sdk/lib/@backend_nest/apps/server/src/controllers/dates/dto/DatesSearchResponse.dto';
-import { EditCategoryForm }       from './EditCategoryForm';
-import { checkFormatDate }        from '@frontend/core/lib/utils/date';
-import { eventParam }             from '../../types';
-import { isRequestError }         from '@frontend/core/lib/utils/isRequestError';
+import { categoryParam, eventParam }             from '../../types';
+import { EventsFetcher } from '../../../components/Fetchers/EventsFetcher';
+import { CategoriesFetcher } from '../../../components/Fetchers/CategoriesFetcher';
+import { DatesFetcher } from '../../../components/Fetchers/DatesFetcher';
+import { EditCategoryView } from './EditCategoryView';
 
 export const EditCategory: React.FC = () => {
-    const { t } = useTranslation('common');
-    const { eventId } = useParams<eventParam>();
+    const { eventId, categoryId } = useParams<eventParam & categoryParam>();
 
-    const [uuid] = React.useState('edit-category-prefetch@' + eventId);
-    const token = useToken();
-
-    const datesResp = useRequest<DatesSearchResponseDto>(
-        {
-            method: 'dates.search',
-            args: [
-                token,
-                {
-                    event: {
-                        $eq: eventId,
-                    },
-                    $sort: [{
-                        $field_name: 'timestamps.event_end',
-                        $order: 'asc',
-                    }]
-                },
-            ],
-            refreshRate: 0,
-        },
-        uuid
-    );
-
-    if (datesResp.response.loading) {
-        return <FullPageLoading/>;
-    }
-
-    if (isRequestError(datesResp)) {
-        return <Error message={t('error_cannot_fetch', { entity: 'date'})} retryLabel={t('retrying_in')} onRefresh={datesResp.force}/>;
-    }
-
-    return <EditCategoryForm
-    dates={datesResp.response.data.dates.map(date => ({
-        id: date.id,
-        name: date.metadata.name,
-        eventBegin: checkFormatDate(date.timestamps.event_begin),
-        eventEnd: checkFormatDate(date.timestamps.event_end),
-    }))}/>;
+    return <EventsFetcher eventId={eventId}>
+        <DatesFetcher eventId={eventId}>
+            <CategoriesFetcher categoryId={categoryId}>
+                <EditCategoryView/>
+            </CategoriesFetcher>
+        </DatesFetcher>
+    </EventsFetcher>
 };
