+++ modules/@frontend_organizer-app/src/screens/UpdateCategories/EditCategory/useCategoryEdition.ts
@@ -1,32 +1,22 @@
-import { CategoriesSearchResponseDto } from '@common/sdk/lib/@backend_nest/apps/server/src/controllers/categories/dto/CategoriesSearchResponse.dto';
 import { CategoriesEditResponseDto } from '@common/sdk/lib/@backend_nest/apps/server/src/controllers/categories/dto/CategoriesEditResponse.dto';
 import { DatesAddCategoryResponseDto } from '@common/sdk/lib/@backend_nest/apps/server/src/controllers/dates/dto/DatesAddCategoryResponse.dto';
 import { useLazyRequest } from '@frontend/core/lib/hooks/useLazyRequest';
 import { useEffect, useState } from 'react';
 import { useDispatch } from 'react-redux';
 import { CategoryPayload, CategoryWithDatesPayload, checkCategory } from '@common/global';
-import { useRequest } from '@frontend/core/lib/hooks/useRequest';
 import { SaleDeltas } from '../../../components/CategoryFields/useCategoryCreationFields';
 import { omit, set } from 'lodash';
 import { PushNotification } from '@frontend/core/lib/redux/ducks/notifications';
 import { useTranslation } from 'react-i18next';
 import './locales';
 import { useDeepEffect } from '@frontend/core/lib/hooks/useDeepEffect';
-import { formatCategoryEntity } from './formatter';
 import { CategoryEntity } from '@common/sdk/lib/@backend_nest/libs/common/src/categories/entities/Category.entity';
 import { useFormik } from 'formik';
 import { useHistory } from 'react-router';
 import { v4 } from 'uuid';
+import { useToken } from '@frontend/core/lib/hooks/useToken';
+import { formatCategoryEntity } from './formatter';
 
-const defaultValues: CategoryWithDatesPayload = {
-    name: '',
-    saleBegin: null,
-    saleEnd: null,
-    seats: null,
-    price: null,
-    currency: 'EUR',
-    dates: [],
-};
 
 interface DateItem {
     id: string;
@@ -34,30 +24,15 @@ interface DateItem {
     eventEnd: Date;
 }
 
-export const useCategoryEdition = (token: string, categoryId: string, dates: DateItem[]) => {
+export const useCategoryEdition = (currCategory: CategoryEntity, dates: DateItem[], refetch: () => void) => {
     const { t } = useTranslation(['edit_category', 'common']);
     const dispatch = useDispatch();
     const history = useHistory();
+    const token = useToken();
 
-    const [fetchUuid] = useState('fetch-category@' + v4());
     const [editUuid] = useState('edit-category@' + v4());
     const [duplicateUuid] = useState('duplicate-category@' + v4());
 
-    const { response: categoryResp, force: forceCategoryReq } = useRequest<CategoriesSearchResponseDto>(
-        {
-            method: 'categories.search',
-            args: [
-                token,
-                {
-                    id: {
-                        $eq: categoryId,
-                    }
-                },
-            ],
-            refreshRate: 0,
-        },
-        fetchUuid
-    );
     const { response: editCategoryResp, lazyRequest: editCategory } =
         useLazyRequest<CategoriesEditResponseDto>('categories.edit', editUuid);
     const { response: duplicateCategoryResp, lazyRequest: duplicateCategory } =
@@ -65,7 +40,6 @@ export const useCategoryEdition = (token: string, categoryId: string, dates: Dat
 
     const [ duplicateDateIds, setDuplicateDateIds ] = useState<string[]>([]);
     const [ relativeSaleDeltas, setRelativeSaleDeltas ] = useState<SaleDeltas>(null);
-    const [ initialValues, setInitialValues ] = useState<CategoryWithDatesPayload>(defaultValues);
 
     const onDuplicate = (saleDeltas: SaleDeltas, dateIdxs: number[]) => {
         setRelativeSaleDeltas(saleDeltas);
@@ -118,7 +92,7 @@ export const useCategoryEdition = (token: string, categoryId: string, dates: Dat
         if (JSON.stringify(category) !== JSON.stringify(formik.initialValues)) {
             editCategory([
                 token,
-                categoryId,
+                currCategory.id,
                 {
                     category: categoryWithoutDate,
                     dates: newDateIds.length > 0 ? newDateIds : undefined,
@@ -153,29 +127,23 @@ export const useCategoryEdition = (token: string, categoryId: string, dates: Dat
         return errors;
     };
 
-    const formik = useFormik({
-        initialValues,
+    const formik = useFormik<CategoryWithDatesPayload>({
+        initialValues: formatCategoryEntity(
+            currCategory,
+            dates
+            .map(({ id }, dateIdx) => {
+                if (currCategory.dates.includes(id)) return dateIdx;
+                return null;
+            })
+            .filter(dateIdx => dateIdx !== null)),
         onSubmit,
         validate,
-        enableReinitialize: true,
     });
 
     useDeepEffect(() => {
-        if (dates && categoryResp.data?.categories[0]) {
-            const category: CategoryEntity = categoryResp.data.categories[0];
-            const dateIdxs: number[] = dates.map(({ id }, dateIdx) => {
-                if (category.dates.includes(id)) return dateIdx;
-                return null;
-            }).filter(dateIdx => dateIdx !== null);
-            setInitialValues(formatCategoryEntity(category, dateIdxs));
-        }
-    }, [dates, categoryResp.data?.categories[0]]);
-
-    useDeepEffect(() => {
-
         if (editCategoryResp.data?.category) {
             dispatch(PushNotification(t('edit_successful'), 'success'));
-            forceCategoryReq();
+            refetch();
             history.replace(history.location.pathname);
         }
     // eslint-disable-next-line
@@ -196,9 +164,6 @@ export const useCategoryEdition = (token: string, categoryId: string, dates: Dat
     }, [duplicateCategoryResp.error]);
 
     return {
-        loading: categoryResp.loading,
-        error: categoryResp.error,
-        forceCategoryReq,
         onDuplicate,
         formik,
     }
