+++ modules/@frontend_organizer-app/src/shared/EventsDrawer/DashboardMenu/View.tsx
@@ -1,72 +1,30 @@
-import React, { useContext, useEffect, useState } from 'react';
-import { useRequest } from '@frontend/core/lib/hooks/useRequest';
-import { EventsSearchResponseDto } from '@common/sdk/lib/@backend_nest/apps/server/src/controllers/events/dto/EventsSearchResponse.dto';
-import { v4 } from 'uuid';
-import { useToken } from  '@frontend/core/lib/hooks/useToken';
-import { UserContext } from '@frontend/core/lib/utils/UserContext';
-import { FullPageLoading, Error, Icon } from '@frontend/flib-react/lib/components';
+import React, { useContext } from 'react';
+import { Icon } from '@frontend/flib-react/lib/components';
 import { useTranslation } from 'react-i18next';
 import './locales';
 import styled from 'styled-components';
 import { useHistory } from 'react-router';
+import { EventsContext } from '../../../components/Fetchers/EventsFetcher';
 
-const NoEventsTitle = styled.span`
-  font-size: 14px;
-  opacity: 0.5;
-  margin-left: 16px;
-`
-
-export const DashboardMenu: React.FC = () => {
+export const DashboardMenuView: React.FC = () => {
     const [ t ] = useTranslation('dashboard_menu');
-    const token = useToken();
-    const [uuid] = useState<string>(v4() + '@events-search');
-    const user = useContext(UserContext);
+    const { events } = useContext(EventsContext);
 
     const history = useHistory();
-
-    const { response: eventsResp, force: forceEvents } = useRequest<EventsSearchResponseDto>({
-        method: 'events.search',
-        args: [
-            token,
-            {
-                owner: {
-                    $eq: user.id
-                },
-                $sort: [{
-                    $field_name: 'updated_at',
-                    $order: 'asc',
-                }]
-            },
-        ],
-        refreshRate: 50,
-    }, uuid);
-
-    useEffect(() => {
-        forceEvents();
-    // eslint-disable-next-line
-    }, []);
-
-    if (eventsResp.loading) {
-        return <FullPageLoading/>;
-    }
-
-    if (eventsResp.error) {
-        return <Error message={t('events_fetch_error')} onRefresh={forceEvents}/>;
-    }
 
     return <DashboardMenuContainer>
         <Title>{t('dashboard_title')}</Title>
         {
-            eventsResp.data.events.length > 0
+            events.length > 0
 
             ?
-            eventsResp.data.events.map(event => (
+            events.map(event => (
                 <Link
                 key={event.id}
                 onClick={() => history.push(`/event/${event.id}`)}>
                     <span>{event.name}</span>
                     <Icon icon={'arrow'} size={'16px'} color={'white'}/>
-                    </Link>
+                </Link>
             ))
 
                 :
@@ -118,4 +76,10 @@ const Link = styled.div`
             opacity: 1;
         }
     }
+`;
+
+const NoEventsTitle = styled.span`
+  font-size: 14px;
+  opacity: 0.5;
+  margin-left: 16px;
 `;
