+++ modules/@frontend_organizer-app/src/shared/EventsDrawer/DashboardMenu/locales.ts
@@ -2,12 +2,10 @@ import i18n from '@frontend/core/lib/utils/i18n';
 
 i18n.addResourceBundle('en', 'dashboard_menu', {
     dashboard_title: 'Events',
-    events_fetch_error: 'Unable to fetch events',
     no_events: 'You have no events yet'
 });
 
 i18n.addResourceBundle('fr', 'dashboard_menu', {
     dashboard_title: 'Évènements',
-    events_fetch_error: 'Impossible de récuperer les évènements',
     no_events: 'Vous n\'avez pas encore d\'évènements'
 });
