+++ modules/@frontend_organizer-app/src/shared/EventsDrawer/DateMenu/CategoriesSubMenu/locales.ts
@@ -2,14 +2,10 @@ import i18n from '@frontend/core/lib/utils/i18n';
 
 i18n.addResourceBundle('en', 'categories_submenu', {
     categories_title: 'Categories',
-    event_fetch_error: 'Unable to fetch event',
-    categories_fetch_error: 'Unable to fetch categories',
     new_category: 'New category',
 });
 
 i18n.addResourceBundle('fr', 'categories_submenu', {
     categories_title: 'Catégories',
-    event_fetch_error: 'Impossible de récuperer l\'évènement',
-    categories_fetch_error: 'Impossible de récuperer les catégories',
     new_category: 'Nouvelle catégorie',
 });
