+++ modules/@frontend_organizer-app/src/shared/EventsDrawer/DateMenu/locales.ts
@@ -5,7 +5,8 @@ i18n.addResourceBundle('en', 'date_menu', {
     dates_and_typology: 'Dates & typology',
     general_infos: 'General informations',
     styles: 'Styles',
-    date_fetch_error: 'Unable to fetch date',
+    from: 'From',
+    to: 'to',
 });
 
 i18n.addResourceBundle('fr', 'date_menu', {
@@ -13,5 +14,6 @@ i18n.addResourceBundle('fr', 'date_menu', {
     dates_and_typology: 'Dates et typologie',
     general_infos: 'Informations générales',
     styles: 'Styles',
-    date_fetch_error: 'Impossible de récuperer la date',
+    from: 'Du',
+    to: 'au',
 });
