+++ modules/@frontend_organizer-app/src/shared/EventsDrawer/EventMenu/CategoriesSubMenu/locales.ts
@@ -2,12 +2,10 @@ import i18n from '@frontend/core/lib/utils/i18n';
 
 i18n.addResourceBundle('en', 'categories_submenu', {
     categories_title: 'Categories',
-    categories_fetch_error: 'Unable to fetch categories',
     new_category: 'New category',
 });
 
 i18n.addResourceBundle('fr', 'categories_submenu', {
     categories_title: 'Catégories',
-    categories_fetch_error: 'Impossible de récuperer les catégories',
     new_category: 'Nouvelle catégorie',
 });
