+++ modules/@frontend_organizer-app/src/shared/EventsDrawer/EventMenu/DatesSubMenu/locales.ts
@@ -2,12 +2,10 @@ import i18n from '@frontend/core/lib/utils/i18n';
 
 i18n.addResourceBundle('en', 'dates_submenu', {
     dates_title: 'Dates',
-    date_fetch_error: 'Unable to fetch date',
     new_date: 'New date',
 });
 
 i18n.addResourceBundle('fr', 'dates_submenu', {
     dates_title: 'Dates',
-    date_fetch_error: 'Impossible de récuperer la date',
     new_date: 'Nouvelle date',
 });
