+++ modules/@frontend_organizer-app/src/shared/EventsDrawer/EventMenu/locales.ts
@@ -7,7 +7,6 @@ i18n.addResourceBundle('en', 'event_menu', {
     published: 'Event visible',
     preview_banner: 'Preview',
     publish_success: 'Event is now published !',
-    event_fetch_error: 'Unable to fetch event',
     no_stripe_interface: 'You need to provide your bank account and other informations first',
     stripe_not_ready: 'Please complete your KYC forms and wait for validation',
     invalid_stripe_interface: 'An error occured, please provide your informations again',
@@ -20,7 +19,6 @@ i18n.addResourceBundle('fr', 'event_menu', {
     published: 'Évènement visible',
     preview_banner: 'Prévisualisation',
     publish_success: 'Évènement publié !',
-    event_fetch_error: 'Impossible de récuperer l\'évènement',
     no_stripe_interface: 'Vous devez fournir un compte bancaire et d\'autres informations pour commencer',
     stripe_not_ready: 'Veuillez completer les formulaires de KYC et attendre leurs validations',
     invalid_stripe_interface: 'Une erreur est survenue, veuillez fournir vos informations à nouveau',
