+++ modules/@frontend_organizer-app/src/shared/EventsDrawer/index.tsx
