+++ modules/@frontend_organizer-app/src/shared/EventsDrawer/locales.ts
